requirements.txt
absl-py==2.1.0
asttokens==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.17.0
fonttools==4.55.6
fsspec==2024.12.0
grpcio==1.70.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.5
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.2.1
numpy==1.24.3
opencv-python==4.8.0.76
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==9.5.0
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==6.1.1
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.13.1
seaborn==0.13.2
sentry-sdk==2.20.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.0.0
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==3.21.0
azure-storage-blob==12.14.1
deep-sort-realtime==1.3.2
flask==2.0.1
Pillow==9.5.0
